fix : #118 홈페이지, 검색페이지 스낵바 버그 수정
