+++ src/pages/search/Search.js
@@ -1,12 +1,12 @@
-import React, {useState, useEffect, useRef, useCallback} from 'react';
-import {useApi} from 'hooks/useApi';
-import {Box, Container, Typography} from '@mui/material';
-import Grid from '@mui/material/Grid';
+import React, { useState, useEffect, useRef, useCallback } from 'react';
+import { useApi } from 'hooks/useApi';
+import { Box, Container, Typography } from '@mui/material';
+import ScrollToTopFab from 'components/ScrollToTopFab';
+import { searchInfoService } from 'api/services/searchInfoService';
+import { useLocation, useNavigate } from 'react-router-dom';
 import SearchBar from 'components/SearchBar';
 import CampingCard from 'components/CampingCard';
-import ScrollToTopFab from 'components/ScrollToTopFab';
-import {searchInfoService} from 'api/services/searchInfoService';
-import {useLocation, useNavigate} from 'react-router-dom';
+import Snackbar from '@mui/material/Snackbar';
 
 function Search() {
     const location = useLocation();
@@ -18,6 +18,10 @@ function Search() {
         city: '',
         keyword: ''
     });
+    const [snackbarMessage, setSnackbarMessage] = useState('');
+    const [snackbarNone, setSnackbarNone] = useState(false);
+    const [snackbarBookmark, setSnackbarBookmark] = useState(false);
+
     const observerRef = useRef(null); // Intersection Observer ref
 
     const {
@@ -103,6 +107,24 @@ function Search() {
         setSearchParams(searchValues);
     };
 
+    const showSnackbarNone = () => {
+        setSnackbarMessage('회원만 이용할 수 있는 기능입니다');
+        setSnackbarNone(true);
+    };
+
+    const showSnackbarBookmark = () => {
+        setSnackbarMessage('찜 상태를 변경하였습니다');
+        setSnackbarBookmark(true);
+    };
+
+    const handleCloseNone = () => {
+        setSnackbarNone(false);
+    };
+
+    const handleCloseBookmark = () => {
+        setSnackbarBookmark(false);
+    };
+
     return (
         <Container maxWidth="lg" sx={{py: 4}}>
             <Box sx={{mb: 4}}>
@@ -141,11 +163,27 @@ function Search() {
                         lineIntro={camp.lineIntro}
                         marked={camp.marked}
                         onClick={() => handleCardClick(camp.campId)}
+                        onShowSnackbarNone={showSnackbarNone}   // 전달
+                        onShowSnackbarBookmark={showSnackbarBookmark} // 전달
                     />
                 ))}
             </Box>
 
             <div id="load-more-trigger" style={{height: '1px'}}/>
+
+            {/* Snackbar 컴포넌트 */}
+            <Snackbar
+                open={snackbarNone}
+                autoHideDuration={6000}
+                onClose={handleCloseNone}
+                message={snackbarMessage}
+            />
+            <Snackbar
+                open={snackbarBookmark}
+                autoHideDuration={6000}
+                onClose={handleCloseBookmark}
+                message={snackbarMessage}
+            />
 
             <ScrollToTopFab/>
         </Container>
